requirements.txt
fastapi
line-bot-sdk
deta